Split the main lib module to into smaller ones
