+++ dacite/__init__.py
@@ -1,0 +1,3 @@
+from dacite.config import Config
+from dacite.core import from_dict
+from dacite.exceptions import *
