+++ dacite/data.py
@@ -1,0 +1,3 @@
+from typing import Dict, Any
+
+Data = Dict[str, Any]
